+++ c/API/Python/TC_ModelFileGenerator_py.c
@@ -20,27 +20,27 @@ static PyObject * pytc_writeModel(PyObject *self, PyObject *args)
 			isList = PyList_Check(pylist);
 			N = isList ? PyList_Size(pylist) : PyTuple_Size (pylist);
 		}
-	
+
 		if (N > 0)
 		{
 			array = malloc( (1+N) * sizeof(void*) );
 			array[N] = 0;
-	
-			for(i=0; i<N; ++i ) 
-			{ 
-				array[i] = isList ? (void*)((int)PyInt_AsLong( PyList_GetItem( pylist, i ) )) : (void*)((int)PyInt_AsLong( PyTuple_GetItem( pylist, i ) ));
-			} 
+
+			for(i=0; i<N; ++i )
+			{
+				array[i] = isList ? (void*)((size_t)PyInt_AsLong( PyList_GetItem( pylist, i ) )) : (void*)((size_t)PyInt_AsLong( PyTuple_GetItem( pylist, i ) ));
+			}
 		}
 	}
 	else
 	{
 		array = tc_allItems();
 	}
-	
+
 	tc_writeModel(s,array);
-	
+
 	free(array);
 	Py_INCREF(Py_None);
-    return Py_None;	
+    return Py_None;
 }
